Finish Entry 012 (Minor changes to 007, 009, 011)
I've also gone ahead and gotten ridden of widowpenalties with the hope
that some of the weird spacing issues get cleared. To be honest, it
should be fine if paragraphs get broken across paragraphs.

I got into sort of a writer's block on this entry mainly because I
didn't feel like writing about all the algebra involved in mathematical
decoding formulas, but I ended up making it out to be worse than it was.
I didn't really have to include all those details because they're quite
messy and the algorithm (it was a nice excuse to learn it) works much
cleaner.