Finish Journal Entry 013
I fixed some of the typesetting in entry 003 regarding vectors and in
the process of making 013 I added a bunch of fun stuff in util so that's
nice. I plan on working more on 007, but to be honest it might take a
bit to get through everything. If I do get to working on that, I might
start on the probability side of things simply because that's the most
fun part for me at least.